+++ changelog.md
@@ -3,3 +3,13 @@
 **Initial release**
 
 This release is the first release of the Fluxins expression parser/evaluator. This release includes a list of default operators and symbols that allows expressions to be expressive.
+
+# v1.0.1
+
+**New Feature**
+
+`fluxins::express` is a new shorthand way to tokenize, parse and evaluate the expression with one function call! 
+
+**Bug Fixes**
+
+Fixed missing include errors, including those that did not cause any errors on my computer.